+++ packages/widget/CHANGELOG.md
@@ -13,4 +13,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Initial release
 
 ## Pending
-
+- Only actor mixin trigger a model change